Allow ThemeChooser to be overriden by a config variable
Sometimes (e.g. for staging/test sites) it's easier to be
able to hard-code what theme the site should use, rather
than rely on the theme itself to correctly recognise the
url. This commit adds a config variable THEME which you
can set to force a particular theme to always be used.