Enable switching between a map and list view of search results
Map view is the default, but shows the list if js is turned off, the
user can switch between the views with a link.

I didn't make switching between map and list javascripted because then
adding deep-linking (and the effects on pagination etc) would be
really hard and require a large extra javascript library + browser
compatibility hassle, so I decided slightly worse is better.