Add a base model for search forms
This implements the ActiveModel interface and backs
a search form, allowing you to use ActiveModel::Validations
and friends to process and perform searches.

This is in app/models so that themes can then override it
as necessary by inheriting from it and filling in the
(currently blank) methods as required.