Use full application urls in all links
This makes sure we link to the full url including the council
reference in search results. To do this in the javascript map
popups I had to add a custom extra bit of data to the
Appliation.as_json so that we can use the the council's
short_name_encoded method.