Zoom map to fit markers after loading. Fixes #139
I played with several options to solve the problem of not being
able to see everything within the default map zoom level
reducing the radius did have some effect, but it still wasn't
optimal on small screens. This however ensures that you always
see everything that your search returned. It has the downside
that on mobile especially you don't see as detailed a view of
the map initially, but this is outweighed I think by the
confusion of not seeing any markers at all. It has the
advantage too of showing you a much *more* detailed view on
larger screens, because it can zoom in as far as possible to
show all the results. I think we should perhaps tweak the
minimum radius down a bit still, but perhaps we can reduce it
to 1 mile or 1/2 a mile instead. Otherwise people don't see
enough results for specific queries.