Add a back to search link to application pages
This adds a link to application pages that takes you back to the
search you came from if the search params are present in the url.

These params are added to the query string of application links in
both the list and map views. They're whitelisted in both controllers
so that no XSS can happen through adding different parameters and
namespaced as search_xxx so that if we ever add url parameters to
application pages, these won't clash.

Closes #171